gallium: interface changes necessary to implement transform feedback
Namely:
- EXT_transform_feedback
- ARB_transform_feedback2
- ARB_transform_feedback_instanced

The old interface was unusable for OpenGL and had to go away.

This interface was originally designed for OpenGL, but additional
changes have been made in order to make st/d3d1x support easier.

The most notable change is the stream-out info must be linked
with a vertex or geometry shader and cannot be set independently.
This is due to limitations of existing hardware (special shader
instructions must be used to write into stream-out buffers),
and it's also how OpenGL works (stream outputs must be specified
prior to linking shaders).

Other than that, each stream output buffer has a "view" into it that
internally maintains the number of bytes which have been written
into it. (one buffer can be bound in several different transform
feedback objects in OpenGL, so we must be able to have several views
around) This counter can optionally be re-set to zero. Also,
the counter can optionally be used to provide the vertex count
for draw_vbo. Note that the counter is supposed to be stored in device
memory and the CPU never gets to know its value.

OpenGL way | Gallium way
------------------------------------
BeginTF    = reset + bind
PauseTF    = unbind
ResumeTF   = bind
EndTF      = unbind
DrawTF     = use pipe_draw_info::count_from_stream_output